Tying dimensions to specific data set
